+++ pybop/models/lithium_ion/echem_base.py
@@ -8,8 +8,8 @@ class EChemBaseModel(BaseModel):
     Overwrites and extends `BaseModel` class for electrochemical PyBaMM models.
     """
 
-    def __init__(self):
-        super().__init__()
+    def __init__(self, name, parameter_set):
+        super().__init__(name, parameter_set)
 
     def _check_params(
         self, inputs=None, parameter_set=None, allow_infeasible_solutions=True